requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.11
openai==0.28.1
pyautogui==0.9.54
pycaw==20230407
requests==2.31.0
python-dotenv==1.0.0
